requirements.txt
openai
jupyter
python-dotenv
pandas
pandas-stubs
scikit-learn
black
mypy
rich